+++ extendr-api/src/wrapper/list.rs
@@ -26,7 +26,7 @@ impl List {
         List::from_values(values)
     }
 
-    /// Wrapper for creating list (VECSXP) objects.
+    /// Wrapper for creating a list (VECSXP) object.
     /// ```
     /// use extendr_api::prelude::*;
     /// test! {
@@ -63,6 +63,33 @@ impl List {
         res.set_names(names).unwrap().as_list().unwrap()
     }
 
+    /// Wrapper for creating a list (VECSXP) object from an existing `HashMap`.
+    /// The `HashMap` is consumed.
+    /// ```
+    /// use extendr_api::prelude::*;
+    /// use std::collections::HashMap;
+    /// test! {
+    ///     let mut map: HashMap<&str, Robj> = HashMap::new();
+    ///     map.insert("a", r!(1));
+    ///     map.insert("b", r!(2));
+    ///
+    ///     let list = List::from_hashmap(map).unwrap();
+    ///     assert_eq!(list.is_list(), true);
+    ///
+    ///     let mut names : Vec<_> = list.names().unwrap().collect();
+    ///     names.sort();
+    ///     assert_eq!(names, vec!["a", "b"]);
+    /// }
+    /// ```
+    pub fn from_hashmap<K>(val: HashMap<K, Robj>) -> Result<Self>
+    where
+        K: Into<String>,
+    {
+        let mut res: Self = Self::from_values(val.iter().map(|(_, v)| v));
+        res.set_names(val.into_iter().map(|(k, _)| k.into()))?;
+        Ok(res)
+    }
+
     /// Return an iterator over the values of this list.
     /// ```
     /// use extendr_api::prelude::*;
@@ -89,6 +116,23 @@ impl List {
         self.names()
             .map(|n| n.zip(self.values()))
             .unwrap_or_else(|| StrIter::new().zip(ListIter::new()))
+    }
+
+    /// Convert a List into a HashMap, consuming the list.
+    ///
+    /// - If an element doesn't have a name, an empty string (i.e. `""`) will be used as the key.
+    /// - If there are some duplicated names (including no name, which will be translated as `""`) of elements, only one of those will be preserved.
+    /// ```
+    /// use extendr_api::prelude::*;
+    /// use std::collections::HashMap;
+    /// test! {
+    ///     let mut robj = list!(a=1, 2);
+    ///     let names_and_values = robj.as_list().unwrap().into_hashmap();
+    ///     assert_eq!(names_and_values, vec![("a", r!(1)), ("", r!(2))].into_iter().collect::<HashMap<_, _>>());
+    /// }
+    /// ```
+    pub fn into_hashmap(self) -> HashMap<&'static str, Robj> {
+        self.iter().collect::<HashMap<&str, Robj>>()
     }
 }